requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==26.1.0
dotenv==0.9.9
Faker==40.1.2
greenlet==3.3.1
iniconfig==2.3.0
packaging==26.0
playwright==1.57.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
typing_extensions==4.15.0
